Refactor presentation folder structure
