+++ i18n/en.pot
@@ -5,16 +5,16 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-07-20T10:55:41.891Z\n"
-"PO-Revision-Date: 2020-07-20T10:55:41.891Z\n"
+"POT-Creation-Date: 2020-08-29T07:47:44.735Z\n"
+"PO-Revision-Date: 2020-08-29T07:47:44.735Z\n"
+
+msgid "Tutorial"
+msgstr ""
 
 msgid "Back"
 msgstr ""
 
 msgid "Help"
-msgstr ""
-
-msgid "Tutorial"
 msgstr ""
 
 msgid "Home"
