+++ src/webapp/themes/dhis2-legacy.theme.ts
